LayoutTests/imported/w3c/web-platform-tests/css/css-contain/content-visibility/content-visibility-anchor-positioning-007-expected.txt

--a1
--a2OO

FAIL CSS Content Visibility: auto not skipped due to being an indirect influence on anchoring assert_approx_equals: #positioned should be anchored because #anchor is on-screen. expected 44 +/- 1 but got -15.109375
